# spring_kotlin/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Sakila API
  version: 1.0.0
  description: API for the Sakila sample database - A DVD rental store management system
  contact:
    name: Pollito
    url: https://pollito.dev
servers:
  - url: http://localhost:8080/api
    description: dev

paths:
  /films:
    get:
      tags:
        - Films
      operationId: findAll
      summary: List all films
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilmListResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /films/{id}:
    get:
      tags:
        - Films
      operationId: findById
      summary: Get film by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilmResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    # Base metadata for all responses
    ResponseMetadata:
      type: object
      properties:
        instance:
          type: string
          description: API endpoint that was called
          example: '/api/something'
        status:
          type: integer
          description: HTTP status code
          example: 200
          minimum: 100
          maximum: 599
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the response was generated
          example: '2026-01-03T17:11:50.826722328Z'
        trace:
          type: string
          description: 32 character lowercase hex String trace identifier. Under unexpected situations it can be '00000000000000000000000000000000'
          example: '9482c151-b417-43ff-9dbb-ee12b84e5d99'
      required:
        - instance
        - timestamp
        - trace
        - status
    # --- Response schemas ---
    # Error response
    Error:
      allOf:
        - $ref: '#/components/schemas/ResponseMetadata'
        - type: object
          description: Standard error response returned when an API request fails
          properties:
            title:
              type: string
              description: HTTP Reason Phrase
              example: 'Not Found'
            detail:
              type: string
              description: Description of the problem
              example: "No static resource for request '/'."
          required:
            - title
    Film:
      type: object
      description: Represents a film available in the Sakila DVD rental catalog
      properties:
        id:
          type: integer
          description: Unique identifier of the film
          example: 42
          minimum: 1
        title:
          type: string
          description: Title of the film
          example: 'ACADEMY DINOSAUR'
          maxLength: 255
        description:
          type: string
          description: Short synopsis of the film
          example: 'An Epic Drama of a Feminist And a Mad Scientist'
          maxLength: 255
        releaseYear:
          type: integer
          description: Year the film was released
          example: 2006
          minimum: 1800
        rating:
          type: string
          description: Motion Picture Association (MPA) rating
          example: 'PG'
        length:
          type: integer
          description: Duration of the film in minutes
          example: 86
          minimum: 0
        language:
          type: string
          description: Language of the film
          example: 'English'
          maxLength: 255
    FilmListResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseMetadata'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Film'
          required:
            - data
    FilmResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseMetadata'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Film'
          required:
            - data